Don't allow for multiple values for one attribute
If you have selected two traits they might end up setting the
same attribute to two different values. This might have you
ending up in an inconsistent state. To handle this Boss Lady
should pull the cord and halt all ongoing work until it has
been sorted out.

The first version is to just show that there's a problem. After
that has been cleaned up and refactored adding the ability
to manually fidget with the attributes seems like a good idea.
(Or possibly just give output suggesting how to create a new
trait, but with BAs and POs being part of the target group this
might not be feasible.)

This commit is an unholy mess, it needs to be cleaned up before being
merged in. But time to catch some zZz.